+++ src/command/list.rs
@@ -15,6 +15,89 @@ pub struct ListOptions {
     /// Group root path
     #[structopt(name = "GROUP", default_value = "")]
     group_path: String,
+    #[structopt(long = "sort", possible_values = &Sorting::variants(), default_value = "numeric-suffix")]
+    sorting: Sorting,
+}
+
+// TODO: Clap 3 will allow this mostly to be derived and documented.
+#[derive(Debug)]
+pub enum Sorting {
+    Lexicographic,
+    NumericSuffix,
+    MaxVox,
+    MaxBlocks,
+    Dtype,
+    Compression,
+}
+
+impl std::str::FromStr for Sorting {
+    type Err = String;
+
+    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
+        #[allow(deprecated, unused_imports)]
+        match s {
+            "lexicographic" => Ok(Self::Lexicographic),
+            "numeric-suffix" => Ok(Self::NumericSuffix),
+            "max-vox" => Ok(Self::MaxVox),
+            "max-blocks" => Ok(Self::MaxBlocks),
+            "dtype" => Ok(Self::Dtype),
+            "compression" => Ok(Self::Compression),
+            _ => Err(format!("valid values: {}", Self::variants().join(", "))),
+        }
+    }
+}
+
+impl Sorting {
+    fn variants() -> &'static [&'static str] {
+        &[
+            "lexicographic",
+            "numeric-suffix",
+            "max-vox",
+            "max-blocks",
+            "dtype",
+            "compression",
+        ]
+    }
+
+    fn sort(
+        &self,
+        a: &(String, DatasetAttributes),
+        b: &(String, DatasetAttributes),
+    ) -> std::cmp::Ordering {
+        match self {
+            Sorting::Lexicographic => Ord::cmp(&a.0, &b.0),
+            Sorting::NumericSuffix => {
+                if let Some((left, right)) =
+                    a.0.split('/')
+                        .zip(b.0.split('/'))
+                        .find(|(last, curr)| last != curr)
+                {
+                    let l = common_prefix_len(left, right);
+                    let left = &left[l..];
+                    let right = &right[l..];
+                    if let (Ok(left), Ok(right)) = (left.parse::<u64>(), right.parse::<u64>()) {
+                        return Ord::cmp(&left, &right);
+                    }
+                }
+
+                Sorting::Lexicographic.sort(a, b)
+            }
+            Sorting::MaxVox => Ord::cmp(&a.1.get_num_elements(), &b.1.get_num_elements()),
+            Sorting::MaxBlocks => Ord::cmp(&a.1.get_num_blocks(), &b.1.get_num_blocks()),
+            Sorting::Dtype => Ord::cmp(
+                &a.1.get_data_type().to_string(),
+                &b.1.get_data_type().to_string(),
+            ),
+            Sorting::Compression => Ord::cmp(
+                &a.1.get_compression().to_string(),
+                &b.1.get_compression().to_string(),
+            ),
+        }
+    }
+}
+
+fn common_prefix_len(a: &str, b: &str) -> usize {
+    a.chars().zip(b.chars()).take_while(|(a, b)| a == b).count()
 }
 
 pub struct ListCommand;
@@ -70,7 +153,10 @@ impl CommandType for ListCommand {
 
         let mut last_path: Option<String> = None;
         let mut format_path = String::new();
-        for (path, attr) in datasets.into_iter().sorted_by(|a, b| Ord::cmp(&a.0, &b.0)) {
+        for (path, attr) in datasets
+            .into_iter()
+            .sorted_by(|a, b| ls_opt.sorting.sort(a, b))
+        {
             let numel = attr.get_num_elements();
             let (numel, prefix) = MetricPrefix::reduce(numel);
             let numblocks = usize::try_from(attr.get_num_blocks()).unwrap();